# Reuse buffer with sync.Pool