While investigating a task #334 to consider `sync.Pool` in our code, I discovered that currently, each invocation of `templateWithValue` creates a new `bytes.Buffer` and discards it. By switching to a `sync.Pool` to reuse and reset these buffers, we can cut down on allocations and reduce overhead.

Here are the benchmark results I gathered using [`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to compare `before_change.txt` (no `sync.Pool`) versus `after_change.txt` (using `sync.Pool`):

```
name                old time/op    new time/op    delta
TemplateWithValue-12    171.5ns ± 4%   145.4ns ± 0%  -15.19% (p=0.000 n=10)

name                old alloc/op   new alloc/op   delta
TemplateWithValue-12    192B ± 0%     80B ± 0%    -58.33% (p=0.000 n=10)

name                old allocs/op  new allocs/op  delta
TemplateWithValue-12      6.0 ± 0%     4.0 ± 0%   -33.33% (p=0.000 n=10)
```
Reusing the buffer with `sync.Pool` is better because it lowers overhead from creating new buffers on each call, reduces memory usage and helps the garbage collector by cutting down on short-lived objects:

- Time per operation improved by about 15%.
- Memory usage dropped by more than 50%.
- Number of allocations reduced from 6 to 4.